refactor: 예약 일자 정보 메서드 이동 (MIKKI-96)
