+++ backend/src/main/java/com/backend/domain/reservation/service/ReservationServiceImpl.java
@@ -1,5 +1,7 @@
 package com.backend.domain.reservation.service;
 
+import java.time.LocalDate;
+
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -10,6 +12,7 @@ import com.backend.domain.reservation.entity.Reservation;
 import com.backend.domain.reservation.exception.ReservationErrorCode;
 import com.backend.domain.reservation.exception.ReservationException;
 import com.backend.domain.reservation.repository.ReservationRepository;
+import com.backend.domain.reservationdate.entity.ReservationDate;
 import com.backend.domain.reservationdate.repository.ReservationDateRepository;
 import com.backend.domain.reservationdate.service.ReservationDateService;
 import com.backend.domain.shipfishingpost.entity.ShipFishingPost;
@@ -17,7 +20,6 @@ import com.backend.domain.shipfishingpost.exception.ShipFishingPostErrorCode;
 import com.backend.domain.shipfishingpost.exception.ShipFishingPostException;
 import com.backend.domain.shipfishingpost.repository.ShipFishingPostRepository;
 
-import jakarta.persistence.EntityManager;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 
@@ -26,7 +28,6 @@ import lombok.extern.slf4j.Slf4j;
 @RequiredArgsConstructor
 public class ReservationServiceImpl implements ReservationService {
 
-	private final EntityManager entityManager;
 	private final ReservationRepository reservationRepository;
 	private final ReservationDateService reservationDateService;
 	private final ReservationDateRepository reservationDateRepository;
@@ -34,8 +35,9 @@ public class ReservationServiceImpl implements ReservationService {
 
 	@Override
 	@Transactional
-	public ReservationResponse.Detail createReservation(ReservationRequest.Reserve requestDto, Long memberId) {
-		log.debug("createReservation transaction start");
+	public ReservationResponse.Detail createReservation(final ReservationRequest.Reserve requestDto,
+		final Long memberId) {
+		log.debug("예약 신청 transaction start");
 		// 선상 낚시 게시글 정보 조회
 		ShipFishingPost shipFishingPost = getShipFishingPostEntity(requestDto.shipFishingPostId());
 
@@ -44,14 +46,14 @@ public class ReservationServiceImpl implements ReservationService {
 			shipFishingPost.getPrice() * requestDto.guestCount());
 
 		// 예약 적용
-		reservationDateService.updateReservationDateWithRemainCount(requestDto.shipFishingPostId(),
-			requestDto.reservationDate(), requestDto.guestCount());
+		updateReservationDateWithRemainCount(requestDto.shipFishingPostId(), requestDto.reservationDate(),
+			requestDto.guestCount());
 
 		// 예약 정보 저장
 		Reservation reservation = reservationRepository.save(
 			ReservationConverter.fromReservationRequest(requestDto, memberId));
 
-		log.debug("createReservation transaction end");
+		log.debug("예약 신청 transaction end");
 		return ReservationConverter.fromReservationResponseDetail(reservation);
 	}
 
@@ -84,6 +86,50 @@ public class ReservationServiceImpl implements ReservationService {
 
 		if (!serverTotalPrice.equals(inputTotalPrice)) {
 			throw new ReservationException(ReservationErrorCode.WRONG_PRICE_VALUE);
+		}
+	}
+
+	/**
+	 * 예약 일자 조회, 예약 가능하면 예약 일자 정보 남은 인원 차감 메서드입니다.
+	 *
+	 * @param shipFishingPostId {@link Long}
+	 * @param reservationDate {@link LocalDate}
+	 * @param guestCount {@link Long}
+	 */
+	private void updateReservationDateWithRemainCount(
+		final Long shipFishingPostId,
+		final LocalDate reservationDate,
+		final Integer guestCount) {
+
+		log.debug("예약 일자 검증 및 업데이트 method start");
+
+		ReservationDate findReservationDate = reservationDateRepository
+			.findByIdWithPessimistic(shipFishingPostId, reservationDate)
+			.orElseThrow(() -> new ReservationException(ReservationErrorCode.RESERVATION_NOT_FOUND));
+
+		verifyReservationDate(findReservationDate, guestCount);
+
+		findReservationDate.remainMinus(guestCount);
+
+		reservationDateRepository.save(findReservationDate);
+
+		log.debug("예약 일자 예약 여부 검증 및 잔여인원 차감 method end");
+	}
+
+	/**
+	 * 선택한 예약 일자의 예약 가능 여부를 검증하는 메서드입니다.
+	 *
+	 * @param reservationDate {@link ReservationDate}
+	 * @param guestCount {@link Integer}
+	 */
+	private void verifyReservationDate(final ReservationDate reservationDate, final int guestCount) {
+
+		if (reservationDate.getIsBan()) {
+			throw new ReservationException(ReservationErrorCode.NOT_AVAILABLE_DATE_RESERVATION);
+		}
+
+		if (reservationDate.getRemainCount() < guestCount) {
+			throw new ReservationException(ReservationErrorCode.NOT_AVAILABLE_END_RESERVATION);
 		}
 	}
 }
